sdk: Move `NullSigner` to `signer` module
