+++ sdk/src/signature.rs
@@ -12,7 +12,7 @@ use std::{
 use thiserror::Error;
 
 // legacy module paths
-pub use crate::signer::{keypair::*, presigner::*, *};
+pub use crate::signer::{keypair::*, null_signer::*, presigner::*, *};
 
 /// Number of bytes in a signature
 pub const SIGNATURE_BYTES: usize = 64;
@@ -110,39 +110,6 @@ impl FromStr for Signature {
         } else {
             Ok(Signature::new(&bytes))
         }
-    }
-}
-
-/// NullSigner - A `Signer` implementation that always produces `Signature::default()`.
-/// Used as a placeholder for absentee signers whose 'Pubkey` is required to construct
-/// the transaction
-#[derive(Clone, Debug, Default)]
-pub struct NullSigner {
-    pubkey: Pubkey,
-}
-
-impl NullSigner {
-    pub fn new(pubkey: &Pubkey) -> Self {
-        Self { pubkey: *pubkey }
-    }
-}
-
-impl Signer for NullSigner {
-    fn try_pubkey(&self) -> Result<Pubkey, SignerError> {
-        Ok(self.pubkey)
-    }
-
-    fn try_sign_message(&self, _message: &[u8]) -> Result<Signature, SignerError> {
-        Ok(Signature::default())
-    }
-}
-
-impl<T> PartialEq<T> for NullSigner
-where
-    T: Signer,
-{
-    fn eq(&self, other: &T) -> bool {
-        self.pubkey == other.pubkey()
     }
 }
